\name{set_cppo} 
\alias{set_cppo} 

\title{Defunct function to set the compiler optimization level} 

\description{This function returns nothing and does nothing except throw a warning. See \url{https://cran.r-project.org/doc/manuals/r-release/R-admin.html#Customizing-package-compilation} for information on customizing the compiler options, but doing so should be unnecessary for normal useage.
} 
 
\usage{
  set_cppo(...)
} 
 
\arguments{
  \item{\dots}{Any input is ignored}
} 


\value{
  An invisible \code{NULL}
}
