\name{read_stan_csv}
\alias{read_stan_csv}
\title{Read CSV files of samples generated by (R)Stan into a \code{stanfit} object}
\description{Create a \code{stanfit} object from the saved CSV files created by Stan or RStan
  to facilitate analysis of samples using RStan.  
} 

\usage{
read_stan_csv(csvfiles)
}

\arguments{
  \item{csvfiles}{A character vector providing CSV file names}
}

\details{
  Stan and RStan could save the samples to CSV files.  This function
  reads the samples and using the comments (beginning with \code{"#"})
  to create a \code{stanfit} object. The model name is derived from 
  the first CSV file. 
} 
\value{
  A \code{stanfit} object (with invalid \code{stanmodel} slot). This \code{stanfit}
  object cannot be used to re-run the sampler. 
} 

\seealso{
  \code{\linkS4class{stanfit}} 
} 
\examples{\dontrun{
## assuming 8schools_1.csv, 8schools_2.csv, 8schools_3.csv are the 
## CSV files generated by either Stan or RStan
f1 <- stan_read_csv(c("8schools_1.csv", "8schools_2.csv", "8schools_3.csv"))
}}
